requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
cffi==2.0.0
click==8.2.1
coverage==7.10.6
cryptography==45.0.7
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
mirakuru==2.6.1
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
port-for==0.7.4
psutil==7.0.0
psycopg==3.2.10
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
pytest-postgresql==7.0.2
pytest-redis==3.1.3
python-dotenv==1.1.1
python-multipart==0.0.20
redis==6.4.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
